Refactoring of local HF backend
