--- requirements_llama2.txt
@@ -1,1 +1,0 @@
-git+https://github.com/facebookresearch/llama.git@main
